+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: nbenhado <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/26 16:19:19 by nbenhado          #+#    #+#             */
-/*   Updated: 2021/11/26 17:39:23 by nbenhado         ###   ########.fr       */
+/*   Updated: 2021/11/26 18:50:36 by nbenhado         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memset( void *pointer, int value, size_t count )
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < count)
